Re-introduce ffishim_library proc macro
There are stubs of code (`free_result`, the allocator setting) that need
to be injected alongside the per-datastructure and per-function
procedural macros.

The approach of inserting these stubs once inside one of the
pre-existing procedural macros does not work in the case where the
exposed API is across multiple crates.

This is because the proc macros are ran once per crate, and the
exclusivity mechanism we use is on a per-instanciation basis. We would
end up with duplicate symbols when linking all crates into one shared
object.

The ffishim_library macro allows the user of the library to define it
only once in the set of crates, preferably in the "main" one.